Wire a context variable through all calls
Unfortunately Golang does not support goroutine local state.
Consequently if we want to pass per-request information through all
the calls we have to do it manually as a method variable. If we are
passing a variable with the per-request identifiers anyways then we
may as well pass the other quantumfs instance specific variables the
same way instead of having to add them individually to the various
classes which need access to them.

We don't use any more stack space than otherwise and the objects
themselves are slightly smaller.