Break NewApi into NewApi and NewApiWithPath
Because all the quantumfs testing is done in parallel we cannot use
chdir() to change to process working directory without interfering
with other tests. Instead, and as a service to other clients which
either have the same restriction or know directly where the api file
is, add a constructor which takes the path directly.