Wire in Inode method setChildAttr
Because of the way the object graph in quantumfs works the attributes
of a file are contained in its parent directory. Thus to change them
we need to call up into our parent.

This implies all the nodes in the workspace subtree knowing their
parent, but we will need that in the near future anyways in order to
push content change notifications upwards to prompt the creation of a
new rootid.