Move test mount into subdirectory
If the quantumfs filesystem under test gets stuck and it's mounted
directly under /tmp, then even running ls in /tmp may get stuck for as
long as the filesystem persists. Avoid this by moving the mounted
filesystem into a subdirectory. We need this eventually anyway when we
start using a persistent cache in some cases.