Abort FUSE connection on panic
If quantumfs panics during a test it is necessary to abort the FUSE
connection or the test process will hang and never exit. In order to
prevent this we need to catch the panic and abort the FUSE connection
before failing the test.

The code here accomplishes this, but only for the case where the FUSE
request is processed on the Server go routine. Unfortunately go-fuse
will use many goroutines and the additional goroutines will crash the
entire test process and not be caught, thus causing the hang.