Un-singletonize qlog
Our testing infrastructure runs many instances of quantumfs in
parallel. In order to allow the test infrastructure to do this and
capture the output from these multiple instances qlog cannot be a
singleton but must instead be unique for each instance of quantumfs.

As part of this refactoring it was necessary to create a Ctx type for
the entire quantumfs package which can carry the request ID beside the
qlog instance which needs to be used.