Move a bunch of Workspaceroot into Directory
The first step in refactoring WorkspaceRoot into just the minimal code
necessary there to differentiate it from Directory and therefore have
recursively supported directories is to move most of the body of
WorkspaceRoot into Directory.

At this point a bunch of tests fail and likely much of the moved code
is not quite correct.