Create self pointer
In golang there isn't any true inheritance, merely embedding. This
implies that, such as in the case of WorkspaceRoot and Directory, you
want a subclass to override a method so the superclass can call it,
that you need to go via an indirect self attribute because all the
methods calls are otherwise statically determined.

That is, if a method of Directory calls dirty(), then it will always
use the Directory copy of that method, never the WorkspaceRoot copy.