Create DirectoryEntry on demand
Previously Directory contained the actual quantumfs.DirectoryEntry
struct. However, this has a single contiguous array of entries which
makes removing entries slow and inefficient. Slow because we'd have to
search through the Entries to find the correct entry. Inefficient
because once we've found it we can create a new slice which doesn't
have that middle element, but since the backing memory is contiguous
we wouldn't release it until the Directory itself was released.

Obviously holding onto an ever growing backing array will cause
problems later. Instead remove the synchronization code and create the
DirectoryEntry from the local maps just before writing to the
datastore.