Add an API command to sync all workspaces.
There doesn't seem to be a filesystem-wide, but also
filesystem-specific sync call. All there is are per-file sync calls
and sync() will call flush() on all open files.

Add an API command to give us the flexibility to sync an entire
quantumfs instance without perturbing the rest of the system.