Convert MultiBlockFile to use an internal map of blocks
This includes renaming the data member to metadata to more clearly
distinguish it from the dataBlocks themselves.

The general theory of operation is that we have a map of all the
blocks which were accessed. Any blocks which have not been accessed
are not in this map at all and therefore won't have been retrieved
from the datastore. We then read and write into these Buffers until we
are asked to sync(), at which point we write out the dirty buffers as
part of generating a new key.