Use a fixed maximum number of DirectoryRecords
In order to ensure that a DirectoryEntry fits into a single block we
need to limit the number of DirectoryRecords recorded. The cheapest
way to do so is to choose a fixed maximum number. Since we currently
do not support variable length filenames this is efficient and easy.

The alternative appears to involve doing many trial marshal attempts,
which would be expensive.