Move the generated Cap'n'Proto files into a new package
One missing feature from Cap'n'Proto in golang is the ability to alias
primitive types. This is an issue because many of the basic types in
the metadata are really just integers of various sizes with different
name and sometimes associated methods.

It has become clear during the conversion to use the new encoding that
without these special type names the existing code will become much
more difficult to read and reason about.

Instead I'm moving the generated encoding files into a new subpackage
and will create type converting wrappers to be used more generally.