Convert all the easy cases to use the new accessors
This is the first phase of converting to use proxy objects between the
daemon code and the Cap'n'Proto encoded metadata blocks. In this
change the basics of the proxy objects have been added and all the
easy accessors been modified to use the method variant instead.

The cases which remain to be handled are converting blocks to metadata
and appending list data to the blocks.