Use hashable key for datastore maps
Cap'n'Proto proxy objects do not produce a unique hash and this
confuses the golang map such that it will not find elements after
they've been inserted. Instead use a constant, hashable value for the
ObjectKey instead.

Right now we produce the human readable string and then hash that.
This isn't especially efficient, but I wasn't able to easily come up
with a better solution. With some benchmarking effort it should be
possible to create an array of the bytes, which will result in the
same hashing performance as we had previously.