Fix off-by-one
When the current DirectoryEntry is full, we need to stop when we hit
the maximum number of children, that way NumEntries() will be
precisely how many we need. Previously we could add one additional
entry, which wouldn't exist in the buffer because space hadn't been
allocated for it. All queries for it would thus return the default
values.