Print workspace name instead of contents
The standard library maps are not concurrency safe and will kill the
process if concurrent accesses are detected while a write is in
progress. It turns out that printing the contents of a map counts as a
read access, obviously enough.

As WorkspaceRoots contain a map we cannot output their contents
without taking a WorkspaceRoot internal lock. Instead of doing that
just skip the full contents entirely.