Move parent into InodeCommon
In order to move entries between directories we need to change its
parent pointer so it knows who its new parent is. Previously all the
Inode types had their own parent attribute which wasn't connected to
any of the others. This made it cumbersome to add a method to change
the parent.

Moving the parent into InodeCommon makes this easy.