Sync srcInode before linking
If srcInode is dirty when it is linked then when the destination
directory calls loadChild_() it will receive out of data data for the
file. Since we are not truly hard linking the new file will
forevermore be out of date.

It is safe to sync srcInode because Directories cannot be hardlinked.