When modifying entry attributes, update ctime and maybe mtime
ctime is the time of the last change on an entry. This could be a
permission or content change. Whenever we modify the attributes of the
entry we will have changed some metadata and therefore should update
the ctime.

mtime is the time of the last contents modification. Because any time
we modify the contents of a directory or file we add an argument here
to also update the mtime is appropriate for the type of attribute
modification being done.