Give all Inodes parents
In order to re-instantiate an Inode we must know its parent, and for
that all the Inodes need a parent.

Thus we pass in parents to ApiInode, WorkspaceRoot and WorkspaceList.
This also requires adding a new Inode method to indicate if the Inode
is a WorkspaceRoot or not, a feat which was previously determined by
testing .parent() against nil.