Activate/deactivate workspace Roots on dirty/publish
If we attempt to activate a previously active workspace we rightfully
panic because we would then have two WorkspaceRoot objects for the
same name. Instead we activate WorkspaceRoots when they are dirtied
and deactivate them after a successful publish when all the changes
have been pushed to the datastore.

This lets us re-instantiate a WorkspaceRoot after it has been
forgotten.