If we fail to forget due to timeout, give up on forgetting
There are two ways we could solve the problem of forgetting. We could
have a long timeout on the first forget and never forget, perhaps on a
per-workspace basis, after the first failure. Or we can have a shorter
timeout for each forget and try to forget for each Forget call.

We cannot have a long forget timeout because then in the failure case
we'll effectively never finish. For example, if we had a one second
timeout and a thousand inodes the kernel knew about, then it'd take
one thousand seconds for the forget phase of unmounting to finish.
This is unacceptable.

A short timeout is also problematic since it will, especially during
tests, be sensitive to system load. 50ms is too short to be reliable,
but much longer runs us into the issue of taking forever to unmount.
This latter causes all the tests to timeout.

Here we take the option of having a longer, but not long, timeout and
giving up on forgetting entirely should we fail once.