Fix BUG 178131 ether panics on Insert in fileStore
Rebased it so that it can be submitted before error handling work.

Root Cause:
There is one instance of fileStore struct(for a qfs daemon) which
implements blobstore.BlobStore. It has a pointer to a rand struct. When 2
fileStore.Insert()s are called in parallel, they use the same random
generator to create random files in writeData().

This leads to data race at:
/usr/lib/golang/src/math/rand/rng.go:233  rng.tap--
/usr/lib/golang/src/math/rand/rng.go:238 rng.feed--

This would explain the out of bounds access at:
/usr/lib/golang/src/math/rand/rng.go:243

Fix:
The random generator is used only to create temporary files in
writeData() in github.com/aristanetworks/ether/filesystem/modify.go

We can replace this with TempFile from io/ioutil package.

Testing:
Created a new test case in unit tests which does 2 parallel inserts
using go routines and ran the test with the   "-race" flag. This exposes
the data race mentioned above. The test case fails in the current
implementation and passes with the fix.

Impact on other bits of ether:
The race condition is not present in cql implementation of
blobstore.BlobStore. Will add the same test case there too.

Also changed make file to run test cases with -race enabled

Change-Id: I6f9617327da69501fac223864f12c0895e442a08