Minor tweaks and scylla workaround.
Incorporated some comments from Krishna T.  about the use of timers in the
utils package.

1. In sendrecv package, changed the upload logic so that there is
one go routine for all the files in the directory. Each directory
still has its own go routine.

Old implementation had one go rountine for each file in a dir
and this was causing the stress command to run out of the number
of open file handles.

2. In filesystem pakage, added a semaphore in the filestore struct,
which is used in filesystem blobstore Insert to limit the number of
files opened by the ether filesystem backend.

3. In cql pakage, added a semaphore in the cqlStore struct, which is used in cql
blobstore Insert to limit the number of insert queries to scyllaDB. This
should be removed when we get a fix from scallaDB.

4. Added a utils package, to add tool and generic blobs of code.
Currently this has an implementation of semaphore using channels. More
to come. The semaphore has a timout field, which is currently set to 60
seconds. When the timeout is reached, it panics.

Change-Id: I16c2b5c62be74306608f485b56fb8cd71c6feb2f