Bug Fix in blobstore integration test case.
Go tests ca be executed in any order and the test cases should handle
that. Changed the TestGet() in blobsstoreIntg_test.go to be complaint
with that. The test case was relying on the fact that the value was
already insereted in the data base using TestInsert().

Also, fixed a hanging lint error.

Change-Id: I860709f812763d25e065d250fb5faf06ade4f6b4