Mount /tmp as tmpfs in non-persistent chroot
Non-persistent chroots are intended for automated builds in ephemeral
workspaces. Consequently we shouldn't care about the persistence of
/tmp in those cases. Moving /tmp off QuantumFS dramatically improves
build speed for packages, such as EosImage, which do complex
operations in /tmp.

All persistent build products should be copied back into an RPM or
installed in the filesystem in any case.