Don't attempt to kick more than once
We can't avoid a deadlock when queueing a dirty inode simply by
releasing the lock before kicking the flusher channel. If we are
queueing the inode because the flusher flushed some other inode which
dirtied its parent. In that situation flushing the queue will be
waiting on queueing a dirty inode, which is waiting on flushing the
queue.

Instead don't try to kick the flusher more than once.