Don't put orphaned Inodes on the dirtyQueue
If an Inode has been instantiated it has been disconnected from the
tree and no modifications to it can make any difference to the
workspace. Thus we shouldn't put it on the dirty queue when modified.

Specifically this solves the problem where a deleted Inode attempts to
sync to its parent, but the parent has been uninstantiated and
forgotten because none of its current children are instantiated to
block it.