Wait for QuantumFS to finish before starting another instance
Unmounting QuantumFS can complete before the final flushing to the
datastore is complete. If we then start a new QuantumFS instance based
upon the same datastore before this flushing is complete the new
instance can read out of date data and act unusually. This is
especially broken when the workspace rootID changes due to the first
QuantumFS still flushing after the second QuantumFS has loaded the
rootID. Since we do not support concurrent workspace modification this
panics as the rootID had been changed remotely.