Convert WorkspaceRoots to return uninstantiated inodes
We don't see this problem with the current code because we always
forget and sync dirty inodes synchronously, but when we asynchronously
clean and forget inodes it is possible that while doing so we must
re-instantiate a WorkspaceRoot. Since Inode re-instantiation is done
with the mapMutex held newWorkspaceRoot calling addUninstantiated()
itself causes a deadlock.