BUG183065: Fix TestPanicFilesystemAbort hang
The issue with TestPanicFilesystemAbort hanging is that it was
panicking before its QuantumFS instance had finished starting because
the system was busy. By making the test wait until its instance has
started it's possible to unmount and recover. If the QuantumFS
instance starts after the test has already failed the recovery runs
before the mount is complete and the test harness then waits forever
waiting for a recovery to complete which was never started against the
filesystem instance.