Don't wait when forgetting a dirty inode
If an Inode is already in the dirty queue when it is forgotten then it
won't be moved to the front of the list in an attempt to allow
children ahead of it to flush themselves first and minimize the number
of writes we do.

However, we should reset the expiry time such that when the flusher
does make it to that Inode it will not wait any additional time.