Approximate directory size by number of entries
Computing the correct size of a directory might be expensive, but
using a fixed value of 1MB is obviously wrong and confusing in many
cases. Instead use an approximation based on a rough maximum per-entry
size.

This isn't perfect, but at least it varies with the number of entries
in a directory.