Reduce log to a warning
There is an intrinsic race between an inode flushing and it being
deleted. One scenario with this race will result in the parent during
the beginning of the flush being different than later in the flush.
This is harmless, it would be preferable to fix the race but that
appears non-trivial.