Close race between Lookup and uninstantiation
This change partially reverts eae2fae185c77ec3bf9e72e89c6c65df39e8b822
(review 391).

With the advent of asynchronous flushing it is possible for an Inode
to be forgotten, queued to be uninstantiated, a Lookup on that Inode
performed and finally for the uninstantiation goroutine to start
uninstantiating that Inode. Fix this race by using the (per workspace)
serialized property of the flusher which prevents races between
uninstantiating an Inode's parent to leave the lookup count entirely
within qfs.lookupCounts.

With that singular source of truth we can then make uninstantiating an
Inode atomic with respect to both Inode instantiation and lookups by
taking all the appropriate locks.