Make childInodes() not return hardlinks
childInodes() is used in the uninstantiation path to determine if all
the children of an inode are uninstantiated and therefore that inode
itself can be uninstantiated. For non-WorkspaceRoot objects this
should not include hardlinks because though those directories contain
a pointer to that inode, they are not its parent. It is thus
inappropriate to keep that directory instantiated because an
non-direct child is still instantiated.

Instead rename childInodes() to directChildInodes() and only return
those children for which the inode is the parent. On Directory this
excludes hardlinks, on WorkspaceRoot this includes them.