Compatibility fix for working with a life QuantumFS filesystem.
It turns out that QFSClient was writing a NUL character after the end of JSON messages, which upsets QuantumFS when it tries to unmarshall these messages.

The existing QFSClient tests relied on this NUL character, so they all passed, although this was bogus.

This change removes the writing of a NUL to the end of messages sent to QuantumFS, and adjusts the tests appropriately.