Calculate nlink on a per-type basis
Previously the nlink count of a directory was not being computed based
upon whether it was a Directory or TypespaceListing or what have you.
Instead it was incorrectly using the same method for all these types.

Move adding the . and .. directory entries into getChildSnapshot()
such that each of the types can calculate the nlink and inode number
as is appropriate for their specific type.