Undo 747e597c8: Test and fix orphaning subtrees
Given the intrinsic issue with using InsertInode with directories
which contain hardlinks, there is no need to allow orphaning subtrees.
That support was added to allow replacing directories instead of
traversing the subtree and deleting all the nodes. If we can only
InsertInode on a single file, requiring an unlink() is reasonable.

Further, it isn't safe to InsertInode over a hardlink as that may
cause the hardlink entry in the WorkspaceRoot to be leaked. Finally,
orphaning in this way created a new class of inode states, disallowing
them reduces the likelihood of related issues in the future.