Do not allow InsertInode on directories
The current structure of QuantumFS does not allow us to use
InsertInode on a directory which contains hardlinks. Doing so may
leave dangling hardlink references or mistakenly point a hardlink to
the content of a hardlink elsewhere in the workspace tree.

Since verifying if that the directory and all its subdirectories does
not contain any hardlinks is expensive, disallow the operation.