Return copies from all datastores and caches
Some filesystem operations, such as extended attribute modifications,
are implemented by interpreting the blocks from the datastore as the
particular metadata type. Some others of these then modify this
metadata and upload the buffer back to the datastore. If the backing
of the buffer from the datastore in the first place is the real
backing, then the data in the datastore will be modified, causing
corruption.

Prevent this for now by always returning a copy of the datastore or
cache data backing.