Do not store clean data in smallFile
Continuing to hold onto the Buffer for a smallFile, even when there is
no dirty data dramatically increases memory usage. The kernel may know
about thousands or tens of thousands of small files and each one could
be consuming up to the maximum block size (1MB) each.

If we can trivially retrieve the data from the datastore there is no
reason for us to store this within the smallFile itself.