Fix a race condition for queueing up a dirty element
The usage of dirtyElementLock was not correct and caused some issues:
1. There was no hierarchy between said lock and dirty queue lock which
in theory could result in a deadlock.
2. There was a race condition that could result in a node being inserted
in the dirty queue more than once.

This patch gets rid of the elementLock and just uses the coarse queue
lock. If this ends up being slow, a better mechanism for queueing up the
dirty nodes can be devised.