Help the user if they don't have capnp installed
We commit the generated metadata.capnp.go file to prevent requiring
people to install the Cap'n'Proto compiler and to ensure all metadata
changes happen consciously.

Sometimes the timestamp of the input file can change, causing a build
failure. Tell the user how to get out of this situation should they
hit it accidentally.