Add password auth to ether.
ether picks up username/password from config file.
The username will need to be setup on the scylla cluster.

I have not made changes to all the config files in this commit.
Just changed the docker cluster_config file in this commit to
show what the change would look like.

Change-Id: I709a80d5a4d9eb186ada3ee080e966f7ed59e006