refresh the hardlinks before the directories
The in-memory hardlinks and dentries will be conflicting in the middle
of the refresh process.

It is simpler to first refresh the hardlinks and then take care of the
dentries as the hardlinks might be looked while walking the directories.
This requires some changes to the directory structure to be able to
handle conflicting information about hardlinks.