Add backoff to walker daemon when walk failures happen.
It is very simple back off.

IF FAILURE IN WALK AT THE END OF AN ITERATION:
	SLEEP 4 HOURS
ELSE
	DO NOT SLEEP

Did not add an eloborate backoff, since we could have failing walks
for certain workspaces for a long time in wsdb and didn't want the
back off period to grown based on that.

Change-Id: If7a12f574856712442644e2baa1721c016add536