Skip marking an inode as accessed if it is orphaned
As we now mark files as accessed for read/update if they are
read/written to and those files can be orphaned at the time (commonly
temporary files), we must support markSelfAccessed() on orphaned
files. On the other hand, we don't want to accidentally call
markAccessed() on an orphaned file as that implies somebody is
accessing a filename from a deleted directory, which I don't believe
is possible.

Therefore, return early in markSelfAccessed() if the inode is
orphaned, but continue to panic in markAccessed() if the inode is
orphaned.