Always use a new inode if there is a type mismatch when refreshing
I was mistaken to assume that I can reuse an old inode id of a file
before refresh for the same file after refresh when the file type
changes.

The major problem is that the old inode might get orphaned and the
kernel will still have references to the old inode with the old inode
id.

If inode id is reused, in these cases the inode id of the old
orphaned file and the inode id of the new entity after refresh
are exactly the same, while they are two completely different files.

Now the code always allocates a new inode id (just like it use to do for
typechange from hardlink before) for any major file type change.

Conflicts:
	daemon/directory_DOWN.go