Unmark deleted then recreated directories
When we have a deleted file which is recreated we count that as a file
which has been updated instead of having all those operations
performed on it. However, there are no operations which would cause a
directory to be marked as updated, thus that configuration of flags
isn't reasonable.

Instead when a directory is deleted and then recreated we will remove
it from the list as if nothing had been done to it. Since deleting a
directories requires first removing all the children of that
directory, they themselves will be listed in the accessed list.
Similarly with files created after the directory has been recreated.