Store strings as hex strings
It appears that Cap'n Proto does not like when golang copies the
struct contents to assign one instance to another value variable.
Unfortunately it appears the json package does this internally
somewhere, which causes the custom UnmarshallJSON() method to
correctly decode the string, but the broader json.Unmarhal() function
does not correctly return that decoded value.

Instead use strings internally and convert to ObjectKey's on the
boundaries.