Do not clear namespace lists on error
If we receive an error from the WorkspaceDB when retrieving the list
of children, it isn't terribly helpful to remove our cached list of
children. If the error is transient we would have caused all the
namespaces to get new inode IDs, and therefore possibly have two
instantiated inodes for the same workspace, for no good reason.

Instead skip the update under the assumption that the error is
transient and will soon correct itself.