Handle PastEndIdx being monotonically increasing
fd2f08a0 subtly changed the meaning of PastEndIdx. Previously it was
the index within the file. That is, it was always modulo the length of
the data memory. After fd2f08a0 it was changed to be a monotonically
increasing number and it became the reader's responsibility to map it
onto the data memory.

Thus all the readers are changed here to perform the modulo mapping
when reading.