Do not leak file handles to the api files
If a test times out, the EndTest() call is going to close the file
handle to the api file and move on to try to umount fuse. However, we do
not join or cancel (in the pthread_join and pthread_cancel sense of the
words) the test goroutines, so they might call getApi() and get a new
file handle to the api file in the meantime, rendering the fuse fs
unmountable.