Retry the syncAll RPC until the inode of interest is forgotten
The hardlink tests assume that a filesystem umount, followed by a
syncAll() should result in the inode of interest getting uninstantiated.
However, a umount results in the kernel only issuing a Forget() call for
the inodes.
The kernel does not wait for the daemon to respond to the forget
call before declaring the umount a success.

An alternative solution would be to drain all pending forget calls
before declaring syncAll a success, but that is inherently racy.