Do not pop the front element as it might not be the processed element
As part of flushInode, the flusher lock is released and re-acquired, a
new element might be pushed to the front of the queue in the meantime
which will be mistakenly removed as soon as flushInode returns.

The bug was introduce in http://gut/review/r/1001 and preserved in the flusher
redesign.