Do not interpret an inodeId missing from the uninstantiated list as being unknown
This assumption is unsafe without the instantiation lock. The inode
might be in the middle of instantiation, during which it gets remvoed
from the list of uninstantiated inodes and then gets added to the list
of inodes.
These two operations are not atomic, and a query between the two might
result in mistakenly declaring the inode as unknown.