add an internal environemnt variable to set CQL CL
By default, the CQL consistency level is quorum.
Under certain circumstances, requirement to work with severly
degraded cluster where we know the data exists, it is useful to
have a way to set the CQL consistency level. This tunable is
not setup in the regular Ether config but available only as
an environment variable and is undocumented.

Change-Id: I017b0793e9b541c9edb6b6ec06a65d1f2abaa8ba