Add a ttl cache
Often QuantumFS will reupload a block which already exists to the data
store to ensure all blocks are correctly in the datastore.. There is
no point in constantly going back to the datastore to confirm the TTL
is up to date. Instead cache that success to reduce network round
trips.