Add locks to Ether WSDB backend
The Ether WorkspaceDB backend doesn't support the atomicity
guarantees the API demands. Until now this hasn't been an issue
because concurrent modification (read-write and write-write) haven't
been supported so there has been no real workspace concurrency. With
the introduction of the WorkspaceDB service this is changing.

Since the wsdbservice should be the only instance making modifications
to the database we can work around the database limitation by locking
at the library level.