Set object type on flush instead at time of change
Since a directory may be flushed before its child having the child
modify its object type, perhaps because the file type changed from a
SmallFile to a MediumFile, can result in an inconsistent metadata tree
state where the object type does not match the type of block the key
points to.

Instead of setting the object type as soon as the file's size becomes
too large, wait until the file itself is flushed to ensure metadata
tree consistency.