Correctly handle the files that are overwritten by a rename
Such a file, shall be orphaned if it has a live inode or should be
removed from the list of uninstantiated inodes.

Testing Done:
Existing unittests and a new one (these fail because of the change made to use elog() for
logging an error).
The TestStressRenames in mut shayan.b215103.4 also passes with this patch.

Reviewed at http://gut/review/r/1208/