qfsclient: add rules to make x86_64 and i686 client RPMs
Add targets to build the 64-bit and 32-bit QuantumFS-client and
QuantumFS-client-devel RPMs

RPM versions are supposed to be one or more dot-separated
decimal numbers. Any additional information can be placed
in the RPM release field.
Add a mechanism for determining the RPM version and release
fields from the git tag. Comments within the Makefile explain
the conversion in more detail.

Add a target to push the RPMs to ToolsV2, as an interim measure,
as the spin-client RPMs are located there.  Ultimately, all of
these RPMs will need to be published to /src/External.