Limit the scope of the inode locks
There is an implicit hierarchy for grabbing the parentLock of
an inode before grabbing its lock.
This hierarchy is violated in some places, so this is not a long-term
fix.

Testing Done:
The tests have been augmented to reproduce this issue. I had to run
the tests for a while to actually hit this issue, but the scenario
is now covered.

Bugs closed: 224715

Reviewed at http://gut/review/r/1217/