Do not check for orphan inodes when dirtying them
The check is not true in the happy path, so there is no performance
optimization in doing it early as the check must be performed before
flushing the inode anyways.

Moreover, the dirty function is called from different call-sites with or
without inode lock and grabbing the parent lock (to check whether the
inode is orphaned) is not safe.