Pass the destination as suffix
In order to create the parent directories if they haven't been
processed yet we need the absolute path to the destination. However,
in order to use InsertInode we need the path relative to the QuantumFS
root. Pass the destination as a suffix and the two different prefixes
to ensure we can handle both cases.