Flush using the initiator's ctx
Most of the time flushing is done because the kicker has triggered it.
However there are times when it is done because a workspace merge
needs to be performed. Debugging is substantially simpler if the
flusher logs with the context of the merge/refresh goroutine in the
latter case instead of the shared flush context.