Handle recreated directories when refreshing a workspace
Refreshing a recreated directory is a special case in which the fileIds
of the source and destination mismatch, but their inode will be reused.

In such cases, the fileId of the localRecord may not exist in the file
map created from remote entries. Update the fileId of the localRecord
immediately to handle this special case.

By design, a directory is never moved as part of refresh. This point is
later asserted as part of refreshing the dentries.