use CFNAME_PREFIX for CF names
This change is for support of Cluster-In-A-Box (CIAB) test environment
so that multiple of these environments can work within the same
keyspace. Each instance of the env, will get it's own tables thus
the keys will not be shared between instances. The keyspace setup
requires cluster admin invtervention. The CIAB framework can
use env vars and the tool "schemar" to setup and teardown the schemas
within the keyspace.

This change allows, the env var CFNAME_PREFIX to name the
blobstore and workspacedb table names.

Testing: unit tests pass
 confirmed that integration tests on dev cluster pass
 confirmed that integration tests in k8 env pass

patchset-3: Refator CQL name check with tests
patchset-2: Fix wsdbstress tool

Change-Id: I720aca5cf0af45aad45a6d81f41c6a8783791d86