Use quantumfs.ObjectKey.String() as the key a map of quantumfs.ObjectKey.
Quantumfs.ObjectKey cannot be used as the key to a map, since the only way to compare 2 quantumfs.ObjectKey(s) is quantumfs.ObjectKey.IsEqual().
So the right way to create a map of quantumfs.ObjectKey(s) we have to use quantumfs.ObjectKey.String() as the key to the map.

Testing: Manual

Change-Id: Ib27efc99ecaa78e6754d2cb5e55cba21d86d6b30