Move skip determination into parent
We need to be able to skip merging individual files. Since there is no
specific mergeInode() style function, move determining if a child
should be skipped into the parent's mergeDirectory() call. This also
changes the input format to no longer accept a leading and trailing /
in order to support a consistent format across files and directories.