Put qloggerdb in a cgroup to force a cg oom-kill and restart
Otherwise, qloggerdb can grow in size and hog all of the physical memory
on a system which will in turn drive quantumfs to a halt.

The issue should be RCA'ed and fixed separately, but the cgroup oom-kill
is a good fail-safe mechanism to recover from these states.

Testing Done:
Manual testing.

Bugs closed: 233958

Reviewed at http://gut/review/r/1355/