Do not overwrite hardlink dentries on load/reload
A directory might get uninstantiated and re-instantiated on the
discretion of the kernel and there was a latent bug that it re-created a
hardlink dentry every time.

Given that the hardlink dentries have non-trivial information (e.g.
ctime) which is used as an input to merge, this would have manifested as
a merge error.

Testing Done:
Existing test now passes

Reviewed at http://gut/review/r/1345/