quantumfs: Update Dockerfile to handle dep-based `make fetch`
The `make fetch` target was revised to use `dep` to fetch and
manage all dependencies under a vendor directory.

Update the Dockerfile to account for this, removing unnecessary
manual git clones that are now handled by `make fetch`

Remove `make` from the package list to quieten a warning from
dnf - it's already installed.
Add `file` to quieten a warning during the cityhash build.