Make all FUSE notification calls asynchronous
It is never allowable to notify FUSE while processing a FUSE
transaction or a deadlock within the kernel will result. It is
difficult to determine precisely which code paths may or may not be
called from a FUSE request and many of them are callable from both.
Move all FUSE notifications to be asynchronous to eliminate this
issue.