Keep a HardlinkDelta object per directory and bubble it up to workspaceroot on sync
This will allow us to always publish a consistent workspaceroot.

There is still remaining work and bugs lurking in this implementation
and corner cases that result in inconsistent workspaceroots.

Specifically, there are three cases about hardlink entry creation,
deletion, and normalization that are not handled correctly yet:

1. If hardlink is created, it will create a hardlink entry
with nlink 2. If that entry gets published in a workspaceroot that has
not yet synced the subtree including the newly created leg, the nlink
of 2 will be wrong.

2. If the last two legs of a hardlink are deleted, but only one subtree
is synced to the workspaceroot, the workspaceroot will contain an
'orphaned' hardlink_leg in one of the subtrees with no corresponding
entry in the hardlink table which is inconsistent.

3. On normalization, a dentry removes the corresponding entry from
hardlink table and inherits it. If the subtree containing the dentry is
not synced yet, the "publishable" workspaceroot will have no knowledge
of the underlying record for the non-normalized dentry in its hardlink
table.

In future reviews, each of these bugs will be reproduced by a
sync/refresh test and fixed.