Treat WSDB errors as final
All the WorkspaceDB errors indicate a situation which will not be
improved by retrying, such as the workspace in question not existing.
Retrying forever in that case simply prevents forward progress.

On the other hand, any error which isn't a WorkspaceDB error likely
comes from either the network or the backend service and is likely to
sort itself out if given sufficient time.