Instantiate a marker file in all workspaces when in low memory mode
This provides users on QuantumFS a way to determine when the
daemon is running in low memory mode and therefore reduced
performance is expected. They can then use this information
to move to the newer instance which is not in low memory
mode.

Testing Done:
New and existing tests. Manual testing to confirm the file appeared
when the daemon has received SIGUSR1 and was readable with cat.

Bugs closed: 229771

Reviewed at http://gut/review/r/1378/